some cleanups in preparing to do 64-bit riff variant. also added generix posix wrappers to make sure all reads, writes, seeks, opens and closes will be 64-bit on any platform.

+++ trunk/src/cnt/riff.h
@@ -30,6 +30,7 @@
 #define RCS_RIFF_H "$RCSfile: riff.h,v $ $Revision: 1762 $"
 
 #include <stdio.h>
+#include <stdint.h>
 
 /* RIFF access return values */
 #define RIFFERR_NONE 0
@@ -52,8 +53,8 @@ typedef unsigned int fourcc_t;
 */
 struct chunk {
   fourcc_t      id;
-  int           start;
-  int           size;
+  uint64_t      start;
+  uint64_t      size;
   struct chunk  *parent;
 };
 typedef struct chunk chunk_t;
